awdur: Refactor to use transforms
I *think* this is the right direction to be moving in.

There already was the transform for inlining project metadata (now
renamed to `ResolveProjectMetadataTransform`), this commit moves the
project construction code into a transform - `BuildProjectsTransform`.

This does make it slightly tricky to share the `project` instance
around, for now the best place for it appears to be the `settings`
on the `document` object.

However, this should make it easier to implement the project browser
directive as that will also need access to the final project
structure.